requirements.txt
#install all these modules before running 
deep-translator==1.11.4
deepdiff==6.7.1
faiss-cpu==1.8.0
googletrans==3.0.0
groq==0.5.0
gTTS==2.5.1
huggingface-cli==0.1
huggingface-hub==0.21.4
langchain==0.1.13
langchain-community==0.0.29
langchain-core==0.1.46
langchain-groq==0.1.3
langchain-text-splitters==0.0.1
langcodes==3.4.0
langdetect==1.0.9
langsmith==0.1.31
language_data==1.2.0
PyAudio==0.2.14
pydantic==2.6.1
pydantic_core==2.16.2
pydub==0.25.1
pygame==2.5.2
pypdf==4.1.0
PyPDF2==3.0.1
pyttsx3==2.90
PyWavelets==1.5.0
semantic-version==2.10.0
sentence-transformers==2.6.0
streamlit==1.32.2
streamlit-chat==0.1.1
tokenizers==0.15.2
torch==2.2.1
transformers==4.39.1
translate==3.6.1

#How to run the code

#1. First install all the required modules
#2. Generate the embeddings
#For embeddings.py
#Use python embeddings.py

#3.Run the chatbot
#For app.py
#Use streamlit run app.py

